Add time zone support and view cleanup
This commit adds basic time zone support to save times by zone in the
database instead of UTC. This should help with summaries not showing
events that are near the day change in some views but not in others.

It also includes some view cleanup, including adding a link to Jira from
the task id.