Add active to Tasks
This commit adds active to Task and modifies the Jira sync to pull the
last 50 user issues modified in the last 2 week. It includes the issue
status, and if it's anything but resolved, the ticket is marked active.
We also added an Task.active scope and updated the project index to only
show active tasks.